+++ packages/elements/src/Matcher/MatcherHelpers.ts
@@ -84,7 +84,7 @@ export function isBaseElementSelectorData(
     "captured",
     "origin",
     "source",
-    "baseSource",
+    "module",
     "isIgnored",
     "isUnknown",
   ]);